Поправил несколько тестов.
