+++ main-service/src/main/java/ru/practicum/controller/PrivateRequestController.java
@@ -1,35 +1,40 @@
 package ru.practicum.controller;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.dto.ParticipationRequestDto;
-import ru.practicum.service.PrivateEventService;
+import ru.practicum.service.ParticipationRequestService;
 
 import java.util.List;
 
+@Slf4j
 @RestController
 @RequestMapping("/users/{userId}/requests")
 @RequiredArgsConstructor
 public class PrivateRequestController {
 
-    private final PrivateEventService privateEventService;
+    private final ParticipationRequestService participationRequestService;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public ParticipationRequestDto createRequest(@PathVariable Long userId, @RequestParam Long eventId) {
-        return privateEventService.createRequest(userId, eventId);
+        log.info("Получен запрос пользователя  с id {} на участие в событии с id {}", userId, eventId);
+        return participationRequestService.createRequest(userId, eventId);
     }
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
     public List<ParticipationRequestDto> getUserRequests(@PathVariable Long userId) {
-        return privateEventService.getUserRequests(userId);
+        log.info("Получен запрос списка всех запросов на участие пользователя с id {}.", userId);
+        return participationRequestService.getUserRequests(userId);
     }
 
     @PatchMapping("/{requestId}/cancel")
     @ResponseStatus(HttpStatus.OK)
     public ParticipationRequestDto cancelRequest(@PathVariable Long userId, @PathVariable Long requestId) {
-        return privateEventService.cancelRequest(userId, requestId);
+        log.info("Получен запрос на отмену запроса c id {} пользователя с id {}&", requestId, userId);
+        return participationRequestService.cancelRequest(userId, requestId);
     }
 }
