+++ main-service/src/main/java/ru/practicum/repository/CategoryRepository.java
@@ -1,10 +1,15 @@
 package ru.practicum.repository;
 
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.model.Category;
+
+import java.util.List;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
     boolean existsByName(String name);
 
     Category getCategoryById(Long catId);
+
+    List<Category> findAllByOrderByIdDesc(Pageable pageable);
 }
